Refactor redundant DBService code to AbstractDBService
